Provide the rendered content in post.content
I firmly believe that post.content should contain the rendered content; it
doesn't make any sense for it to be otherwise -- how often are you going to
want the pre-rendered content as opposed to the ready-to-read version.  For
anyone who *does* want the raw version in their Liquid templates, there's
now post.raw_content.