New preserve_mtime config option and logic
Caching can be greatly optimised if mtimes on files are consistent.  This
option makes it possible to set the mtime of the files to that of the mtime
of the source file.  This doesn't take into account the possibility that
layouts or includes have changed, but for the general case it works well.