Add support for posts as subdirectories
In addition to supporting posts as single files, allow posts to be the
index file of a directory named for the post.  Require the same naming
conventions for the directories as was required for the files.

This feature was requested in Issue 511.  I foresee the main use for
this feature as allowing users to more easily manage content associated
with a single post by placing it all in the same directory and without
the need to duplicate the post name in the filename and containing
directory.

Functionality for dealing with the post content may be provided by
plugins or future commits.  This patch only provides support for post
file naming.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>