Update pygments.rb to 0.3.2
When using jekyll with pygments.rb 0.2.x, the development server did not
work with automatic reloading enabled. Thanks to git-bisect, I traced
this problem to b2a1d61, which is the commit that introduced pygments.rb
support. I'm not sure of the specific bug in pygments.rb that caused this
issue, but t is resolved by using a newer version of pygments.rb.

Signed-off-by: Michael Dippery <michael@monkey-robot.com>