add ability to set site-wide layout defaults
If no specific layout is set on the post or page,
ask the `Site` class for a default layout for the
given path and type (page or post).

The `Site` class loops through an array of layout_defaults
hashes set in configuration and for each decides whether it
applies to the given post or page and whether it overrides
the other applying defaults (by means of sub-path overrides
super path).

It then returns the name of the default layout or nil.