Update Jekyll gemspec.
The master gemspec has gotten a tad out of date; this updates the
manifest, notably, but also the version number, to bring it in line with
v1-stable.  I'm not sure this is precisely the right way for the latter
to work; @parkr, would you like to comment on this?