Windows fix for 1.4.3 issue 1948 + 1946
There are several things afoot in the Windows 1948 issue.
1) the drive is expanded into the File.expand_path when set relative to
"/" - I believe the idea was to circumvent escaping project root using
../ but these are taken care of in url.rb, and escaping can still be
done using the encoded %2E%2E/
2) to alleviate the escaping url.rb has been modifiend to replace any
occurence of %2[eE] with a . which then later will be picked up in case
it is part of ..
3) The inclusion of '+' in filenames would be unescaped into ' ' which
is a rather poor directory name in the '+/plus+in' test. Thus the
destination method needs to escape plusses to %2B before they are
unescaped.
4) the approximately same code for the method destination was found in
two subclasses (page.rb and post.rb) of convertible.rb. The method has
now been pulled up into convertible.rb