Header links to pages which have a title
Header used to add link to all pages of the site, even those which do not have any title, creating empty anchors on the page like `<a class="page-link" href="/feed.xml"></a>`. Those were non click-able and used space on the page due to margins.

Now only displays pages with a title.