Fix path case change in sanitized_path on Windows
Under Windows the use of File.expand_path can change the case of the path
being sanitized if there is a file or folder with the same name (but
different case) in the root of the drive containing the working directory.

This replaces the call to expand_path with regular expressions to achieve
the same result without the side effects.

Also add some extra tests to ensure that sanitize_path is working as
intended and for the case change behaviour under Windows.