Return nil when comparing a `Post` with a non-`Post`.
In Ruby < 2.2, `==` will swallow any exceptions raised by `<=>` and
return nil.

In Ruby > 2.2, `==` will surface any exceptions raised by `<=>`.

In Ruby 2.2, `==` will show an annoying warning message if `<=>` raises
an exception, which drowns out most of Jekyll's output.

This patch double-checks that the `date` and `slug` methods exist on
whatever you're comparing a `Post` to, and returns `nil` if they
don't. This keeps compatibility with Ruby 2.2+.