replace broken link; add other Mac-specific info
The link for 'use Xcode to install Command Line Tools' seems to be broken. I don't know what it contained, but I have replaced it with a short piece of text which describes how to do the same thing. I have also added a bit more advice which was helpful for me.

FYI, the existing Mac advice on this page includes using sudo. I have been told that using sudo in these commands is not advisable, but since I have no direct knowledge of the pros and cons of that (in this specific case), I am not completely removing that advice, only adding the alternative of not using sudo.