Add space and Rus-lang chars support in filenames
I needed this feature in my project, as forced to use Russian-language filenames in some cases. I think it's useful not only for me and I think that in the future should enter at least limited support for UTF filenames for other languages. If the project team will agree with this, I can even prepare the necessary regular expression for a part of national alphabets.