DOC: Describe --allow-hash-href arg for htmlproof
It is very confusing to new users (it was to me at least) when you initialise a fresh jekyll site with `jekyll new PATH`, add the tests to `.travis.yml`, and then find that they don't pass.

It took a while for me to discover that this is because the initial site includes links to `#`, and there is the `--allow-hash-href` argument which can be supplied to `htmlproof` to allow this behaviour.

This commit adds a description of the `--allow-hash-href` argument and the need to use it for a fresh initialisation of jekyll to the documentation.