Add 5 errors for config, render, doctor errors
- UnhealthySiteError is thrown when a site doesn’t pass all the `jekyll
doctor` tests
- InvalidConfigurationError is thrown if a configuration file returns
an invalid value (e.g. false) or when the configuration has set a value
that requires human intervention to change (e.g. using maruku and
setting to kramdown would cause problems so site needs to be audited
for upgrade to work properly
- InvalidFileFormatError is thrown when a non-.toml/.y(a)ml file is
given as a config file to be read in. It may be used later for invalid
source files.
- FileNotFoundError is thrown when a configuration file could not be
found. It will be used later for other errors, including trying to
include a file that’s not there.
- IncorrectDependencyError is thrown when a dependency is activated
that is not a valid version, e.g. a version of rouge that is < 1.3.0 is
activated.

Specs are also added.