Add collection support to the "post_ur"l tag
These changes add very simplistic support for collections to the post_url tag, making it easier to link to files that reside within collections (in response to issue #2540). 

Functionally, these changes cause the post_url tag to split the tag input parameter on all spaces. If the array generated by the split operation is exactly two elements long, the post_url tag will use the first element as the name of the collection and the second element as the document to find in the collection. If the array is not exactly two elements long, post_url will act as normal.

This is my first time playing around with the Jekyll source, so I'd be happy to have some feedback; I'm sure there's a better way to parse out multiple tag arguments, but split(" ") was the best I could find right now. It might also be reasonable to instead parse the incoming URL to see if it starts with a known collection name, and go from there.