Add a Gemfile to the site template.
This adds a Gemfile to the site template and sets the `ruby` to the currently running Ruby version.  This will "cheat the system" in that if a user is using an older version of Ruby, it will install an older version of dependencies.  Which according to: https://github.com/bundler/bundler-features/issues/120#issuecomment-214514847 should in theory work.  Only @indirect can confirm or deny this will actually happen and perhaps let us know if this is currently available within stable.