Add option to fail a build with front matter syntax errors
This commit adds a config named `strict_front_matter`. This configuration causes a build to fail if there is a syntax error in a page's front matter.

Prior to this commit, if a page had an error in its front matter the build would print a warning, ignore the file, and continue. This caused problems where pages would have a syntax error and the build would not fail, and instead build an incomplete site.

This commit adds the option which tells a build to fail when there is a syntax error reading / parsing a page's front matter. This is done by setting the `strict_front_matter` config or using the `--strict_front_matter` CLI option.

This commit also adds a test case for this behavior and updates the documentation.