Add Utils::Internet.connected? to determine whether host machine has internet connection.
Many plugins make calls out into the Internet for data. In order to facilitate a pattern
for gracefully falling back when there is no Internet connection for the user, this module
and module function are added. Plugins which access the Internet should decide whether there
is a path for them to gracefully fallback to some non-Internet default when the host computer
is not connected.