Update quickstart.md
I think the "~ $" that starts each line is unnecessary and on fist glance a potential (yet minor) point of confusion. Everyones terminals  come in different shapes and sizes, so I think sticking to the commands is often the best bet. Especially with such a basic and brief, section.

"~/myblog $ bundle exec jekyll serve" in particular can be confusing for people following along. Given it's a quick start, many users will just be doing a copy and paste, or just typing out what they see. On a glance you could be easily mistaken in thinking "~/myblog" is there so run the command from this location and is not in fact the current directory with the command following after the "$".

It's a very minor change, but I think it will read better.