Front matter defaults without default_proc
This is a PoC of a system which shims accessing data in front of front
matter defaults to eliminate the need for default_proc.

Use of Hash#default_proc has caused some confusion and frustration,
especially when trying to marshal and unmarshal Documents, or Pages, or
any other Hash which has an attached default_proc. Unfortunately, procs
cannot be marshaled.

All the default_proc was doing was asking the site's frontmatter
defaults which values applied to the given object and passing them
along. Here, we place a custom class in front of the Hash data and reach
out to the front matter defaults when applicable for the given object.

An alternative method would be to strip and re-add the default_proc
before and after marshaling, respectively.