Provide a better default hash for tracking liquid stats
This avoids doing the `||=` check for an existing value when setting
values for the per-file stats hash in the liquid renderer. It also gives
us the flexibility to add new numerical values to this hash later
without having to do the same `||=` check as we can just know that the
starting value will be zero.