Add LiveReload functionality to Jekyll
Invoke using jekyll serve --livereload. The Serve command will inject
some necessary JavaScript links into pages. When loaded in a browser,
this JavaScript opens a connection over port 35729 using WebSockets to
an EventMachine reactor listening to that port in a separate thread.

When Jekyll finishes rendering a page, --livereload hooks create a
JSON message with the relative URL of the refreshed page. This message
is then pushed to the browser over the WebSockets connection. The
client-side JavaScript then reloads the page if the URL in the message
matches the URL of the current page.

Implements #4644