Switch to an actively-maintained TOML parser.
We have an ancient version of the `toml` gem which appears to have
been abandoned for years and isn't compliant with TOML 0.4.0, the
current state of the art.

The `tomlrb` and `toml-rb` gems are both (more) actively maintained
and kept up-to-date; switching to either gives us TOML 0.4.0 support.
This patch switches us to `tomlrb`.

(See also: #6540 #6541)