Add ordinal_date filter
It is normal to write things like _posted on 12th February 2018" in the
blog post. Currently to achieve this, a code similar to this one is
needed:

```
{% assign d = page.date | date: "%-d" %}
{% case d %}
{% when '1' or '21' or '31' %}
{{ d }}st
{% when '2' or '22' %}
{{ d }}nd
{% when '3' or '23' %}
{{ d }}rd
{% else %}{{ d }}th
{% endcase %}
{{ page.date | date: "%B" }}
{{ page.date | date: "%Y" }}
```

I think the case is common enough to introduce it in Jekyll, by adding
some parameters to the `date_to_string` and `date_to_long_string`
methods.

It supports both UK and US styles. UK is the default.