Move Bundler Instructions to Tutorial
Based on discussion in PR #6828, move the instructions for setting up a
project with Bundler to a tutorial. This is a better place for it since
the installation instructions are targeted primarily at new users, and
it can be confusing to list several different installation methods -
particularly more complex ones.