Prefer RubyInstaller over WSL
While WSL is still a valuable option to work with jekyll, things have
changed a lot at RubyInstaller, so that it has become super easy to
setup jekyll native on Windows. With the "Ruby+Devkit" versions
of RubyInstaller it's enough to click through the installer, to be able
to issue "gem install jekyll" or "github-pages" successfully.

This also updates WSL instructions a little bit.