Added example for block tags, plus other tweaks
The documentation as it stood left me in many tangles when I first came to Jekyll. I know there is plenty of support available elsewhere, but I feel this needs just a little more.

I have added a whole example for block tags, and have given a small amount of info on how parameters are passed around. I've also updated the examples to use the Tags/Filters module, which matches the convention of Jekyll's own tags.

Finally, I've fixed the examples to show the parameters being non-destructively manipulated. This is to help new developers avoid horrible gotchas that can occur if they do mutate the parameters, particularly when their custom tags are placed inside Liquid forloops.