Respect SOURCE_DATE_EPOCH environment variable (if set) over Time.now
Whilst working on the Reproducible Builds effort [0], we noticed
that jekyll builds sites that are not reproducible.

This is because site.time uses the current time of day and thus
anything then shipping a Jekyll site (eg. as documentation) will
therefore not be reproducible.

This commit prefers the value in the SOURCE_DATE_EPOCH environment
variable [1] if it is set and config.time is also not set.

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/