Add note to "env at build" about multi-config
There are two ways that allow to work with Jekyll and multiple environments. The section "Specifying a Jekyll environment at build time" is the best place to learn about those. But until now, this section only described one of the two options. This change add a note-area to the page that references the other option. I also removed the paragraph about the "avoid editing the config file", since the second option that is now described is the better solution for this use case.