Add some minor improvements to image loading
This removes the loading animated image from the "showcases" view
and sets a default height for each image to avoid page "jumping"
on image loading.

Added a small "spacer" gif in order to have a src to load on render
and a background image. When the image loads has a small fade in
effect.