add clarification on downcase
Based on my testing, a front matter needs to be added before the downcase will have an effect. Without the `---\n---` at the top of the index.html doc, the row as indicated in line 64 appears in its raw form when rendered in the browser. 

As soon as I added the Front Matters, the site rendered as expected. I have added row 72 to clarify that for other beginners like myself.