Update 07-assets.md
After I started the step by step tutorial, I got a bit confused at the point where to add the main.scss file. First we need to create a Sass file at /assets/css/styles.scss, so at this moment we are focused on the folder 'assets'. And then next we should create a styles.css under _sass, for me was not 100% clear if I should create this folder insede assets/_sass or /assets/css/_sass