add support for journaling
When maintaining a journal it doesn't make sense to name the file slug.
Just create one entry per day.

Actually coming up with a slug name before the post if even written
doesn't make sense. Half of the time I end-up renaming the slug, which
is cumbersome to do.

This change makes it possible to use vimwiki to create new post entries.