Implement a `pop_item` filter
This is just a proxy for Array#pop in Ruby. We're using nil as a
indicator that a number was not passed. If we use 1 as the default,
Array#pop always returns an array. If the number `1` is passed in, then
we just let Array#pop always return an array.