Add .gitattributes
.gitattributes sets such configuration as line-endings; cross-platform
projects require a .gitattributes to control how files are checked out
on various platforms.

The base project will check out in native line ending (committing to the
repository in a standard normalized format of Unix style line endings).
The test resources will be placed on-disk without any translation, so
that tests can expect a particular format.