Clarify `bundle config` in Bundler tutorial
I wrote this Bundler tutorial back in 2018. At that time, running
`bundle install --path vendor/bundle` was the correct way to install
gems to a custom directory with Bundler. Since then, Bundler released
[version 2.1](https://github.com/rubygems/bundler/blob/master/CHANGELOG.md#210pre1-august-28-2019),
which deprecates breaking changes for Bundler 3. One of these changes is
the deprecation of the `install --path` option, which has been replaced
by `bundle config set path foo`. The tutorial was updated for this in #8048.

This PR builds upon clarifies that change. We use `--local` to avoid
system-wide user preference changes (which is implied by the info box
beneath). In addition, I wanted to clarify the reasons you might want to
follow this approach, after reviewing this tutorial with a fresh set of
eyes.