Added hidden flag --config
After spending hours trying to figure out a solution to my issue, I stumbled upon [a SO answer](https://stackoverflow.com/a/48292461/205696) that mentions the `--config` flag, which is paramount to solving my use-case.

Then I found it is hidden away in the documentation, as the second section under *Configuration Options* that starts out with explaining options in `_config.yml`.

I expected all CLI commands, options and arguments to documented under *Command Line Usage* or via `jekyll --help`, like most other terminal programs. In lieu of that I have tried my best to ease the discovery of the CLI commands (in their entirety).

It is unfortunate that CLI usage is split in the documentation between CLI usage and `_config.yml` options but to fix that, I think a bigger restructure of the docs are necessary.