Deploy docs from 68497a4d4b075797689264adad45bb74a1004117
Dan Nemenyi: Echo PATH to .zshenv for newer macOS's (#8483)

* Add PATH to Zsh for newer macOSs

Apple switched the default shell to Zsh with Catalina, so updating the PATH in .bash_profile will not have effect for relevant users, and is increasingly now the legacy approach.

This commit adds instructions to echo the PATH to .zshenv, the Zsh equivalent of .bash_profile, as well as to check which version of macOS the user is using.

* Update PATH in local macOS documentation