Add the Hombrew gems directory to the PATH
When you install Ruby with Homebrew, and you install a gem the regular
way, without any flags, it gets installed in
`/usr/local/lib/ruby/gems/2.7.0/bin`. If that directory is not in the
`PATH`, you will get "command not found" when trying to use that gem.

This is a common issue people run into when they type
`gem install jekyll` instead of `gem install jekyll --user-install`.
Do a search for "jekyll: command not found" in your favorite search
engine, in GitHub, Stack Overflow, etc.

Updating the documentation to also include the gems directory will
prevent this issue and will make people happier.

Ideally, if this were my project, I would replace pretty much this
entire document with 3 steps:

1. Link to a reliable, repeatable script that can set up a proper
Ruby environment with a Ruby manager

2. Quit and restart Terminal

3. `gem install jekyll`

It should, and can be this easy.

Here are examples of scripts I have written:
https://github.com/monfresh/install-ruby-on-macos
https://github.com/monfresh/laptop/

Discourse is another example of a Ruby project that provides a script:
https://meta.discourse.org/t/beginners-guide-to-install-discourse-on-macos-for-development/15772