Update action configuration to use current defaults
The default GitHub branch when creating a repository is now "main" and no longer "master", but included a comment for those with older repositories.
- https://github.com/github/renaming
- https://www.zdnet.com/article/github-to-replace-master-with-main-starting-next-month/

The most recent version of the jekyll-action GitHub Action, 2.1, has changed the name of the environment variable from `JEKYLL_PAT` to `token`.
- https://github.com/marketplace/actions/jekyll-actions#deprecation

Adding these two changes should make it easier for anyone setting up the GitHub Action, especially on a recently created repository. It will also help using the latest version of the action to bring it into line with current documentation.