Update filters.rb
When iterating over objects and passing them to the 'slugify' filter, results in an error if they aren't already a string.

For example: (debug with --trace)
>         2: from C:/Ruby27-x64/lib/ruby/gems/2.7.0/gems/jekyll-3.9.0/lib/jekyll/filters.rb:63:in `slugify'
>        1: from C:/Ruby27-x64/lib/ruby/gems/2.7.0/gems/jekyll-3.9.0/lib/jekyll/utils.rb:219:in `slugify'
> jekyll-3.9.0/lib/jekyll/utils.rb:367:in `replace_character_sequence_with_hyphen': undefined method `gsub' for 2021:Integer (NoMethodError)

In this case I'm iterating over post.tags and one of the tags is "2021". This object is recognized as an integer and will not 'slugify'. Changing it to a string in 'filters.rb' before it's passed to 'utils.rb' solve this issue.