Improve the "Markdown Options" Docs
It seems the Markdown Options docs have become out of date, as noted in
https://github.com/jekyll/jekyll/issues/8593. The docs most likely
became out of date when Kramdown was updated to v2, as described in
these release notes:
https://jekyllrb.com/news/2020/08/05/jekyll-3-9-0-released/.

Initially, I intended to fix the specific issue about `gfm_quirks` noted
in #8593, but after digging a little deeper to figure out the correct
way to use that option, it became apparent that other parts of the
Markdown Options Jekyll docs were also outdated and the whole page
needed to be updated to be more accurate and more useful for modern
versions of Jekyll.

So, with that in mind, here's an overview of the changes I'm proposing:

* Adjust the page headings. `Kramdown` should be an `h2` and the
  Kramdown sub-headings should be `h3`.
* Document modern config for the GFM parser via `input: GFM`.
* Link to the CodeRay docs and provide an example of the correct
  `syntax_highlighter` config with `syntax_highlighter_opts`.
* Link to the options supported by Kramdown and provide an example of
  their usage.
* Remove the list of all the relatively advanced Kramdown options
  supported by Jekyll, directing users instead to the Kramdown docs.

At a high level, this is a move away from attempting to document every
supported Kramdown option on this page and a move toward linking to
up-to-date documentation for the various gems we support. The old docs
were outdated, and not as useful as the docs provided by individual gems
because the Jekyll docs didn't provide thorough descriptions as the
[Kramdown docs](https://kramdown.gettalong.org/options.html) do. I think
my proposed change captures a good balance of showing example usage in
Jekyll while still relying on the gems we depend on to provide thorough
descriptions of the supported options.