Improve GitHub Flavored Markdown Docs
I recently merged #8681 to improve the Jekyll docs for markdown options.
While researching that, I noted that the `kramdown-parser-gfm` gem is
required by default in Jekyll, but I didn't dig further as to why.

After #8681 merged, I was curious about why `kramdown-parser-gfm` is
required by default, and I followed up on digging into it. As it turns
out, this GFM parser is required by default because Jekyll uses GFM by
default. This is clear from the [default config
docs](https://jekyllrb.com/docs/configuration/default/) and the code:

https://github.com/jekyll/jekyll/blob/76517175e700d80706c9139989053f1c53d9b956/lib/jekyll/configuration.rb#L67-L72

Although this is outlined in the default configuration docs, other parts
of the docs don't do a good job making it clear that GFM is actually the
default behavior. My recent docs change in #8681 made this problem worse
since I used a bunch of config in my examples that's actually just
default config.

In this PR, I'm addressing that problem by modifying my update to the
Markdown Options docs to clearly note that GFM is the default parser.
I'm also updating a tutorial that implied GFM isn't the default. I've
searched this repo for "GFM" and I'm pretty confident I've found all the
problematic docs.