Implements sort_order attribute for collections
Picks up #3735 where @parkr suggested a simple implementation of @goodtouch's suggestion to allow collection sorting to be more configurable.

This is a minimal implementation that bolts onto the existing approach. However the collection is sorted, the user can simply reverse this. This is particularly useful for date ordering, where ascendant and descendant are equally valid sortings depending on the circumstance.