Fix random test failures in TestExcerpt #to_liquid.
Failures can occur when the random seed orders TestFilters tests before
TestExcerpt and the tests are run with a system time zone that's has a
non-zero offset on 2013-07-22.

The TestFilters setup changes ENV["TZ"] to "UTC" but doesn't reset it.

The TestExcerpt#to_liquid test fails because @excerpt.to_liquid["date"]
returns a UTC time and Time.parse("2013-07-22") returns a time in the
system time zone:

  Failure: TestExcerpt#test_: An extracted excerpt #to_liquid should
  contain the proper page data to mimic the post liquid.
  [jekyll/test/test_excerpt.rb:103] Minitest::Assertion:
  Expected: 2013-07-22 00:00:00 +0000
    Actual: 2013-07-22 00:00:00 +0100

Add a teardown to TestFilters to reset ENV["TZ"].