removed wrong description of root directory
In this document, they describe the root directory as the directory `where your config file is`. This is true in most cases, when the user hasn't changed the `source` [configuration option](https://jekyllrb.com/docs/configuration/options/). However, when `source` is set to another directory, then the root directory of files is in another directory from the directory of the config file.

Please let me know if the text I have removed adds clarity, or if I should make a reference to the `source` option from the configuration