Deploy docs from 300668ed7fe229f9ef7e448a2d05ba5d165e2a22
Ashwin Maroli: Lock `psych` to v4.x

Psych 5 has stopped bundling `libyaml` and expects it to be installed on the host system prior
to being invoked.
Since we don't have a direct dependency on the Psych gem (it gets included in the gem bundle
as a dependency of the `rdoc` gem), lock psych gem to v4.x instead of installing `libyaml` in our
development / CI environment.