Use FileUtils#remove_entry_secure in StaticFile#write
Sometimes the `dest_path` in Jekyll::StaticFile#write is a directory. When this happens, `FileUtils.rm(dest_path)` will raise an exception.

Using `FileUtils.remove_entry_secure(dest_path)` instead will not raise an exception and remove (aka “unlink”) file system entries if the file `dest_path` points to is a directory or a non-directory file.