Update min Ruby to 3.1.3 and update macOS docs
As reported in #9613, the latest version of bundler requires Ruby version 3.0.0 or greater, so when people follow the quickstart instructions while using an older Ruby version, they will get an error.

As such, I think the minimum Ruby version should be the oldest version that hasn't yet reached end of life, which is 3.1.6. Furthermore, Ruby versions older than 2.6.10 cannot be safely installed natively on Apple Silicon Macs.

As for the macOS docs, I updated the supported macOS versions to be in line with what Apple and Homebrew support, and it's no longer necessary to install xz separately. It gets installed automatically when you install ruby-install.