Update minimum Ruby version to 3.1.6
As reported in #9613, the latest version of Bundler requires Ruby version 3.0.0 or greater, so when people follow the quickstart instructions while using an older Ruby version, they will get an error.

As such, I think the minimum Ruby version should be the oldest version that hasn't yet reached end of life, which is 3.1.6 currently. Furthermore, Ruby versions older than 2.6.10 cannot be safely installed natively on Apple Silicon Macs.

As for the macOS docs, I updated the supported macOS versions to be in line with what Apple and Homebrew support. I also removed `xz` from the  prerequisites that need to be installed with Homebrew since it gets installed automatically when you install ruby-install.