Deploy docs from 5861d3ab284cb321ff987d8dcb2358a0647b39c8
Moncef Belyamani: Update current Ruby version to 3.3.5 (#9614)

This is a documentation change to update the current version of Ruby to the
latest as of today: 3.3.5

As for the macOS docs, I updated the supported macOS versions to be in line with what Homebrew supports. I also removed `xz` from the  prerequisites that need to be installed with Homebrew since it gets installed automatically when you install ruby-install.

Co-authored-by: Moncef Belyamani <git@rubyonmac.dev>