Deploy docs from 1b617d72815a0c7fc90214a9dee7ae19cb27010e
Molly Exten: Updated note on Gemfiles in 10-deployment.md (#9805)

This is a 🔦 documentation change.

I've adjusted the documentation to include a note that Gemfile is
already installed if the reader has completed step 1 in this tutorial. I
thought it was kind of confusing to see instructions on adding a Gemfile
when I had already done that step in the first part of the tutorial.

---------

Co-authored-by: Matt Rogers <mattr-@github.com>