Update WDM (0.1.1 → 0.2.0) in the Windows docs
Windows Directory Monitor (WDM) is a library to monitor directories on Windows. It's mostly implemented in C.
However, WDM 0.1.1 can no longer compile with Ruby 3.3+.

On 2024-08, a maintainer of rubyinstaller updated WDM for newer Ruby versions in https://github.com/Maher4Ever/wdm/pull/29, and publish WDM 0.2.0.
The new version declared to be compatible with Ruby 2.5+.

This PR update the WDM version from 0.1.1 to 0.2.0 in the Jekyll docs for Windows.