Adds one entry on the Directory Structure page to include a brief description of the _plugins directory.
Since it is a relevant component of the directory structure, it
should have a brief mention there. The funcionality of plugins (and
of the usage of this folder) is more extensively covered in the Plugins
section of the documentation, and it is also referenced in the
description of the directory.

Resolves: #9894